Remove criterion benchmarks
