+++ benches/path_tree.rs
@@ -1,11 +1,15 @@
-//! Benches sourced from `matchit` (MIT AND BSD-3-Clause)
-//! <https://github.com/ibraheemdev/matchit/blob/v0.8.6/benches/bench.rs>
+#![allow(clippy::incompatible_msrv)]
 
-use criterion::black_box;
+//! Benches sourced from `path-tree` (MIT OR Apache-2.0)
+//! <https://github.com/viz-rs/path-tree/blob/v0.8.1/benches/bench.rs>
+
+use std::hint::black_box;
+
 use divan::AllocProfiler;
+use path_tree_routes::paths;
 use smallvec::SmallVec;
 
-pub mod matchit_routes;
+pub mod path_tree_routes;
 
 #[global_allocator]
 static ALLOC: AllocProfiler = AllocProfiler::system();
@@ -17,12 +21,12 @@ fn main() {
 #[divan::bench(name = "wayfind")]
 fn wayfind(bencher: divan::Bencher<'_, '_>) {
     let mut router = wayfind::Router::new();
-    for route in routes!(brackets) {
-        router.insert(route, true).unwrap();
+    for (index, route) in routes!(brackets).iter().enumerate() {
+        router.insert(route, index).unwrap();
     }
 
     bencher.bench(|| {
-        for path in black_box(routes!(literal)) {
+        for path in black_box(paths()) {
             let _output = black_box(router.search(black_box(path)).unwrap());
         }
     });
@@ -31,12 +35,12 @@ fn wayfind(bencher: divan::Bencher<'_, '_>) {
 #[divan::bench(name = "wayfind (parameters)")]
 fn wayfind_params(bencher: divan::Bencher<'_, '_>) {
     let mut router = wayfind::Router::new();
-    for route in routes!(brackets) {
-        router.insert(route, true).unwrap();
+    for (index, route) in routes!(brackets).iter().enumerate() {
+        router.insert(route, index).unwrap();
     }
 
     bencher.bench(|| {
-        for path in black_box(routes!(literal)) {
+        for path in black_box(paths()) {
             let output = black_box(router.search(black_box(path)).unwrap());
             let _parameters: SmallVec<[(&str, &str); 4]> =
                 black_box(output.parameters.iter().map(|p| (p.0, p.1)).collect());
@@ -53,7 +57,7 @@ fn actix_router(bencher: divan::Bencher<'_, '_>) {
     let router = router.finish();
 
     bencher.bench(|| {
-        for path in black_box(routes!(literal)) {
+        for path in black_box(paths()) {
             let mut path = actix_router::Path::new(path);
             let _output = black_box(router.recognize(black_box(&mut path)).unwrap());
         }
@@ -69,7 +73,7 @@ fn actix_router_params(bencher: divan::Bencher<'_, '_>) {
     let router = router.finish();
 
     bencher.bench(|| {
-        for path in black_box(routes!(literal)) {
+        for path in black_box(paths()) {
             let mut path = actix_router::Path::new(path);
             let _output = black_box(router.recognize(black_box(&mut path)).unwrap());
             let _parameters: SmallVec<[(&str, &str); 4]> =
@@ -81,12 +85,12 @@ fn actix_router_params(bencher: divan::Bencher<'_, '_>) {
 #[divan::bench(name = "matchit")]
 fn matchit(bencher: divan::Bencher<'_, '_>) {
     let mut router = matchit::Router::new();
-    for route in routes!(brackets) {
-        router.insert(route, true).unwrap();
+    for (index, route) in routes!(brackets).iter().enumerate() {
+        router.insert(*route, index).unwrap();
     }
 
     bencher.bench(|| {
-        for path in black_box(routes!(literal)) {
+        for path in black_box(paths()) {
             let _output = black_box(router.at(black_box(path)).unwrap());
         }
     });
@@ -95,12 +99,12 @@ fn matchit(bencher: divan::Bencher<'_, '_>) {
 #[divan::bench(name = "matchit (parameters)")]
 fn matchit_params(bencher: divan::Bencher<'_, '_>) {
     let mut router = matchit::Router::new();
-    for route in routes!(brackets) {
-        router.insert(route, true).unwrap();
+    for (index, route) in routes!(brackets).iter().enumerate() {
+        router.insert(*route, index).unwrap();
     }
 
     bencher.bench(|| {
-        for path in black_box(routes!(literal)) {
+        for path in black_box(paths()) {
             let output = black_box(router.at(black_box(path)).unwrap());
             let _parameters: SmallVec<[(&str, &str); 4]> =
                 black_box(output.params.iter().map(|p| (p.0, p.1)).collect());
@@ -117,7 +121,7 @@ fn ntex_router(bencher: divan::Bencher<'_, '_>) {
     let router = router.finish();
 
     bencher.bench(|| {
-        for path in black_box(routes!(literal)) {
+        for path in black_box(paths()) {
             let mut path = ntex_router::Path::new(path);
             let _output = router.recognize(&mut path).unwrap();
         }
@@ -133,9 +137,9 @@ fn ntex_router_params(bencher: divan::Bencher<'_, '_>) {
     let router = router.finish();
 
     bencher.bench(|| {
-        for path in black_box(routes!(literal)) {
+        for path in black_box(paths()) {
             let mut path = ntex_router::Path::new(path);
-            let _output = router.recognize(&mut path).unwrap();
+            router.recognize(&mut path).unwrap();
             let _parameters: SmallVec<[(&str, &str); 4]> =
                 black_box(path.iter().map(|p| (p.0, p.1)).collect());
         }
@@ -145,12 +149,12 @@ fn ntex_router_params(bencher: divan::Bencher<'_, '_>) {
 #[divan::bench(name = "path-tree")]
 fn path_tree(bencher: divan::Bencher<'_, '_>) {
     let mut router = path_tree::PathTree::new();
-    for route in routes!(colon) {
-        let _ = router.insert(route, true);
+    for (index, route) in routes!(colon).iter().enumerate() {
+        let _ = router.insert(route, index);
     }
 
     bencher.bench(|| {
-        for path in black_box(routes!(literal)) {
+        for path in black_box(paths()) {
             let _output = black_box(router.find(path).unwrap());
         }
     });
@@ -159,12 +163,12 @@ fn path_tree(bencher: divan::Bencher<'_, '_>) {
 #[divan::bench(name = "path-tree (parameters)")]
 fn path_tree_params(bencher: divan::Bencher<'_, '_>) {
     let mut router = path_tree::PathTree::new();
-    for route in routes!(colon) {
-        let _ = router.insert(route, true);
+    for (index, route) in routes!(colon).iter().enumerate() {
+        let _ = router.insert(route, index);
     }
 
     bencher.bench(|| {
-        for path in black_box(routes!(literal)) {
+        for path in black_box(paths()) {
             let output = router.find(path).unwrap();
             let _parameters: SmallVec<[(&str, &str); 4]> =
                 black_box(output.1.params_iter().map(|p| (p.0, p.1)).collect());
@@ -174,13 +178,13 @@ fn path_tree_params(bencher: divan::Bencher<'_, '_>) {
 
 #[divan::bench(name = "route-recognizer")]
 fn route_recognizer(bencher: divan::Bencher<'_, '_>) {
-    let mut router = route_recognizer::Router::new();
-    for route in routes!(colon) {
-        router.add(route, true);
+    let mut router = route_recognizer::Router::<usize>::new();
+    for (index, route) in routes!(colon).iter().enumerate() {
+        router.add(route, index);
     }
 
     bencher.bench(|| {
-        for path in black_box(routes!(literal)) {
+        for path in black_box(paths()) {
             let _output = black_box(router.recognize(path).unwrap());
         }
     });
@@ -188,13 +192,13 @@ fn route_recognizer(bencher: divan::Bencher<'_, '_>) {
 
 #[divan::bench(name = "route-recognizer (parameters)")]
 fn route_recognizer_params(bencher: divan::Bencher<'_, '_>) {
-    let mut router = route_recognizer::Router::new();
-    for route in routes!(colon) {
-        router.add(route, true);
+    let mut router = route_recognizer::Router::<usize>::new();
+    for (index, route) in routes!(colon).iter().enumerate() {
+        router.add(route, index);
     }
 
     bencher.bench(|| {
-        for path in black_box(routes!(literal)) {
+        for path in black_box(paths()) {
             let output = router.recognize(path).unwrap();
             let _parameters: SmallVec<[(&str, &str); 4]> =
                 black_box(output.params().iter().map(|p| (p.0, p.1)).collect());
@@ -205,12 +209,12 @@ fn route_recognizer_params(bencher: divan::Bencher<'_, '_>) {
 #[divan::bench(name = "xitca-router")]
 fn xitca_router(bencher: divan::Bencher<'_, '_>) {
     let mut router = xitca_router::Router::new();
-    for route in routes!(colon) {
-        router.insert(route, true).unwrap();
+    for (index, route) in routes!(colon).iter().enumerate() {
+        router.insert(*route, index).unwrap();
     }
 
     bencher.bench(|| {
-        for path in black_box(routes!(literal)) {
+        for path in black_box(paths()) {
             let _output = black_box(router.at(path).unwrap());
         }
     });
@@ -219,12 +223,12 @@ fn xitca_router(bencher: divan::Bencher<'_, '_>) {
 #[divan::bench(name = "xitca-router (parameters)")]
 fn xitca_router_params(bencher: divan::Bencher<'_, '_>) {
     let mut router = xitca_router::Router::new();
-    for route in routes!(colon) {
-        router.insert(route, true).unwrap();
+    for (index, route) in routes!(colon).iter().enumerate() {
+        router.insert(*route, index).unwrap();
     }
 
     bencher.bench(|| {
-        for path in black_box(routes!(literal)) {
+        for path in black_box(paths()) {
             let output = router.at(path).unwrap();
             let _parameters: SmallVec<[(&str, &str); 4]> =
                 black_box(output.params.iter().map(|p| (p.0, p.1)).collect());
